+++ dataflow/local_run/local_run.py
@@ -33,11 +33,25 @@ FILEGROUPS = ['10_meteo']
 #               '40_chambers_ghg']
 # # FILEGROUPS = ['10_meteo']
 
+# # CH-FOR
+# SITE = 'ch-for'
+# DATATYPE = 'raw'
+# # DATATYPE='processing'
+# # FILEGROUPS = ['10_meteo']
+# FILEGROUPS = ['10_meteo']
+
 # # CH-FRU
 # SITE = 'ch-fru'
 # DATATYPE = 'raw'
 # DATATYPE='processing'
 # FILEGROUPS = ['10_meteo', '13_meteo_pressure']
+# FILEGROUPS = ['10_meteo']
+
+# # CH-HON
+# SITE = 'ch-hon'
+# DATATYPE = 'raw'
+# # DATATYPE='processing'
+# # FILEGROUPS = ['10_meteo', '13_meteo_pressure']
 # FILEGROUPS = ['10_meteo']
 
 # # CH-LAE
@@ -77,30 +91,30 @@ FILEGROUPS = ['10_meteo']
 
 # Common xxx
 ACCESS = 'server'
-DIRCONF = r'F:\Sync\luhk_work\20 - CODING\22 - POET\configs'
-# YEAR = 2017
-MONTH = None
-# MONTH = 6
+DIRCONF = r'L:\Sync\luhk_work\20 - CODING\22 - POET\configs'
+YEAR = 2010
+# MONTH = None
+MONTH = 4
 FILELIMIT = 0
 # FILELIMIT = 10
 NEWESTFILES = 0
-# TESTUPLOAD = True
-TESTUPLOAD = False
-# N_ROWS = 5  # Only upload x number of rows of each file
-N_ROWS = None
+TESTUPLOAD = True
+# TESTUPLOAD = False
+N_ROWS = 5  # Only upload x number of rows of each file
+# N_ROWS = None
 INGEST = True
-# INGEST = False
+# INGEST = False  # If False, VarScanner will run normally, but no data uploaded (faster)
 
 # For parallel processing of months or years
-# MONTHS = range(1, 13, 1)
-YEARS = range(2005, 2021, 1)
+# MONTHS = range(4, 8, 1)
+# YEARS = range(2016, 2017, 1)
 # filegroup = '10_meteo'
 
 kwargs = dict(site=SITE,
               datatype=DATATYPE,
               access=ACCESS,
               dirconf=DIRCONF,
-              # year=YEAR,
+              year=YEAR,
               # filegroup=filegroup,
               month=MONTH,
               filelimit=FILELIMIT,
@@ -117,13 +131,13 @@ if __name__ == '__main__':
     # Run years and all filegroups in parallel
     # for month in MONTHS:
     #     kwargs['month'] = month
-    for year in YEARS:
-        kwargs['year'] = year
-        for filegroup in FILEGROUPS:
-            kwargs['filegroup'] = filegroup
-            p = multiprocessing.Process(target=run_dataflow, kwargs=kwargs)
-            p.start()
-            processes.append(p)
+    # for year in YEARS:
+    #     kwargs['year'] = year
+    for filegroup in FILEGROUPS:
+        kwargs['filegroup'] = filegroup
+        p = multiprocessing.Process(target=run_dataflow, kwargs=kwargs)
+        p.start()
+        processes.append(p)
 
     for p in processes:
         p.join()